[Reconstruction] Action

+++ src/main/java/indi/etern/checkIn/action/setting/save/SaveVerificationSetting.java
@@ -1,32 +1,33 @@
 package indi.etern.checkIn.action.setting.save;
 
-import indi.etern.checkIn.action.TransactionalAction;
+import indi.etern.checkIn.action.BaseAction1;
+import indi.etern.checkIn.action.MessageOutput;
 import indi.etern.checkIn.action.interfaces.Action;
+import indi.etern.checkIn.action.interfaces.ExecuteContext;
+import indi.etern.checkIn.action.interfaces.InputData;
 import indi.etern.checkIn.entities.setting.verification.VerificationRule;
 import indi.etern.checkIn.service.dao.VerificationRuleService;
+import org.springframework.transaction.annotation.Transactional;
 
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 
 @Action("saveVerificationSetting")
-public class SaveVerificationSetting extends TransactionalAction {
-    List<Map<String, Object>> dataList;
+public class SaveVerificationSetting extends BaseAction1<SaveVerificationSetting.Input, MessageOutput> {
+    public record Input(List<Map<String, Object>> data) implements InputData {}
     final VerificationRuleService verificationRuleService;
     
     public SaveVerificationSetting(VerificationRuleService verificationRuleService) {
         this.verificationRuleService = verificationRuleService;
     }
     
+    @Transactional
     @Override
-    public String requiredPermissionName() {
-        return "save verification setting";
-    }
-    
-    @Override
-    protected Optional<LinkedHashMap<String,Object>> doAction() throws Exception {
+    public void execute(ExecuteContext<Input, MessageOutput> context) {
+        context.requirePermission("save verification setting");
+        final Input input = context.getInput();
         verificationRuleService.deleteAll();
+        final List<Map<String, Object>> dataList = input.data;
         for (Map<String, Object> data : dataList) {
             Map<String, Object> property = (Map<String, Object>) data.get("property");
             List<String> trace = (List<String>) property.get("trace");
@@ -45,12 +46,6 @@ public class SaveVerificationSetting extends TransactionalAction {
                     .build();
             verificationRuleService.save(verificationRule);
         }
-        return Optional.ofNullable(getSuccessMap());
-    }
-    
-    @SuppressWarnings("unchecked")
-    @Override
-    public void initData(Map<String, Object> dataMap) {
-        dataList = (List<Map<String, Object>>) dataMap.get("data");
+        context.resolve(MessageOutput.success("Verification setting saved"));
     }
 }
